added user auth, favorites products,

+++ model/cart.js
@@ -1,9 +1,10 @@
 import connectionPromise from "./connexion.js";
 
-export const getCartModel = async () =>{
+export const getCartModel = async (userId) =>{
+    console.log(userId);
     let connection  = await connectionPromise;
     let results = await connection.all(`SELECT * FROM Products p INNER JOIN 
-                                        cart c ON p.id = c.productId`);
+                                        cart c ON p.id = c.productId where c.user_id = ?`,userId);
     return results;
 }
 
@@ -20,41 +21,49 @@ export const getCartProductById = async (id) => {
     return result;
   };
 
-export const deleteCartProduct = async (id) =>{
-    console.log(id);
+export const deleteCartProduct = async (id,userId) =>{
     let connection  = await connectionPromise;
-    await connection.run("DELETE FROM cart WHERE productId = ?",id);
+    await connection.run("DELETE FROM cart WHERE productId = ? and user_id=?",id,userId);
     let result = await connection.all(`SELECT * FROM Products p INNER JOIN
-                                        cart c ON p.id = c.productId`);
+                                        cart c ON p.id = c.productId 
+                                        where user_id = ?`,userId);
+
 
     return result;
 }
 
-export const incrementCartModel = async (id) =>{
+export const incrementCartModel = async (id,userId) =>{
     let connection  = await connectionPromise;
-    await connection.run("UPDATE cart SET quantity = quantity + 1 WHERE productId = ?",id);
+    await connection.run(`UPDATE cart SET quantity = quantity + 1 
+                            WHERE productId = ? and user_id= ?`,id,userId);
 
     let result = await connection.all(`SELECT * FROM Products p INNER JOIN
-                                        cart c ON p.id = c.productId`);
+                                        cart c ON p.id = c.productId
+                                        where user_id = ?`,userId);
     return result;
 }
 
-export const decrementCartModel = async (id) =>{
+export const decrementCartModel = async (id,userId) =>{
     let connection  = await connectionPromise;
-    await connection.run("UPDATE cart SET quantity = quantity - 1 WHERE productId = ?",id);
+    await connection.run(`UPDATE cart SET quantity = quantity - 1 
+                            WHERE productId = ? and user_id = ?`,id,userId);
 
     let result = await connection.all(`SELECT * FROM Products p INNER JOIN
-                                        cart c ON p.id = c.productId`);
+                                        cart c ON p.id = c.productId
+                                        where user_id = ?`,userId);
     return result;
 }
 
-export const getSubTotal = async () =>{
+export const getSubTotal = async (userId) =>{
     let connection  = await connectionPromise;
+    console.log(userId);
     let result = await connection.get(`SELECT SUM(CAST(SUBSTR(p.price, 2) AS DECIMAL) * c.quantity ) as subTotal 
                                             FROM Products p 
                                             INNER JOIN cart c 
-                                            ON p.id = c.productId;
-                                            `);
+                                            ON p.id = c.productId
+                                            where user_id = ?`,userId);
+                                            ;
+                                          
     return result;
 }
